requirements.txt
tk
Pillow
customtkinter